refactor: deploy modular Recon architecture (9 services) - High Cohesion & Low Coupling
